rename base styles and add comment
